Motors working
I put this on my quad but the RX is not reading fast enough. I’m
switching to a mega in the meantime as it has more interrupts. Once I
get it all figured out I’ll switch back to the Uno as communication
will be coming over serial (from ROS).