Added more functionality
Was able to get this to work in the .ino by change extension to “.h”.

Added starting of currently unused “arrayStretchTo”, for use with
making arm count dynamic also for use with pwm max buffer.

Added JacksonSimpleFilter which is filtering by averaging amounts in an
array. Every time a value is shifted in, the last value is shifted out.